Create new base class for entities using udfs
Move code related to udf handling from class Artifact to the new class
Udf. Make both Artifact and ExtensionContext inherit from Udf, and in
future any class using udfs.
